--- src/main/java/com/example/gatherplan/appointment/dto/RegionDto.java
@@ -1,11 +1,0 @@
-package com.example.gatherplan.appointment.dto;
-
-import lombok.*;
-
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RegionDto {
-    private String address;
-}
