+++ src/main/java/com/example/gatherplan/controller/mapper/RegionVoMapper.java
@@ -1,7 +1,7 @@
 package com.example.gatherplan.controller.mapper;
 
-import com.example.gatherplan.appointment.dto.*;
 import com.example.gatherplan.controller.vo.appointment.*;
+import com.example.gatherplan.region.dto.*;
 import org.mapstruct.*;
 
 @Mapper(
