+++ src/main/java/com/example/gatherplan/region/dto/RegionDto.java
@@ -1,4 +1,4 @@
-package com.example.gatherplan.appointment.dto;
+package com.example.gatherplan.region.dto;
 
 import lombok.*;
 
@@ -6,6 +6,6 @@ import lombok.*;
 @Builder
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RegionReqDto {
-    private String keyword;
+public class RegionDto {
+    private String address;
 }
