+++ src/main/java/com/example/gatherplan/region/repository/CustomRegionRepository.java
@@ -1,6 +1,6 @@
-package com.example.gatherplan.appointment.repository;
+package com.example.gatherplan.region.repository;
 
-import com.example.gatherplan.appointment.repository.entity.Region;
+import com.example.gatherplan.region.repository.entity.Region;
 
 import java.util.Optional;
