+++ src/main/java/com/example/gatherplan/region/service/RegionService.java
@@ -1,6 +1,7 @@
-package com.example.gatherplan.appointment.service;
+package com.example.gatherplan.region.service;
 
-import com.example.gatherplan.appointment.dto.*;
+import com.example.gatherplan.appointment.dto.CSVRowDto;
+import com.example.gatherplan.region.dto.*;
 
 import java.util.List;
 
@@ -10,4 +11,6 @@ public interface RegionService {
     List<KeywordPlaceRespDto> searchKeywordPlace(KeywordPlaceReqDto reqDto);
 
     List<DailyWeatherRespDto> searchDailyWeather(DailyWeatherReqDto reqDto);
+
+    void saveFromCSV(List<CSVRowDto> rows);
 }
